Inventory has specified size
Inventory fills itself with empty slots on initialization and pickups will not react if inventory is full. Previously this script would append to the _content array when new items were collided with, now, after looking for partially full slots, it looks for slots with the 'empty' item to replace.

Known issue: if pickups can be partially picked up, the pickup on the ground is not reduced by the amount picked up by the player, creating a duplication of those items.